Implemented the simple logout button functionality , also improved the toast messages for authentication
